Info route shows basic config details from Admin
Laying the groundwork for the UI - no styling / layout just basic tags
and tables for now. I'm planning on coming back at the end to add
styling or even load in a boostrap template.

Process of loading the info was painstaking, decided not to use an ember
data model as API does not seem to allow making changes to this data nor
does it adhere to REST conventions as a resource. I am not thrilled how
the model hook for the info route is handled but it'll work for now.

Tests have taken a backseat for now, TODO mock an endpoint and assert
values showed up on the rendered page (integration style)