Adding things to the adapter for application
The process of adapting the data coming in from Kong API was fairly
painless but PATCHing an update to the Kong API ran into a few speed
bumps.

First off deep in the source of ember there is an aggressive
`adapter.ajaxOptions` hook that does a JSON.stringify if two conditions
are met:

 1. there is data
 2. it's not a GET request

This posed a problem as the server was expecting www-form-urlencoded
data and the stringify was adding quotes to the end of my encoded form
data. So far the only way i've seen to avoid that stringify check is to
override the ajaxOptions, let it play through, and then go back and
slice the quotes off.  I'm fairly certain this will cause issues when I
get to the point where I want to add an API record but we'll cross that
bridge when we get there.